Augmented add-task, refresh-tasklist and clear-tasklist with test vectors

+++ model/requests/delete-all-cards.go
@@ -28,7 +28,7 @@ var DeleteAllCardsRequest = Request{
 						Name: "controller",
 						Type: "uint32",
 					},
-					Value: uint32(405419896),
+					Value: 405419896,
 				},
 			},
 			Expected: []byte{
